bootstrap 3 and jquery
I just switched bootstrap version from 4 to 3 because some of the commands on W3school.com  do not work on version 4. I did so because i thought it will be easier for us to learn quickly from the site but if you guys don't like it, please feel free to switch it back.

Also, I added jquery into js folder as some of the commands in bootstrap require jquery.